move some files to actually match the readme
